+++ packages/kit/src/runtime/server/page/render.js
@@ -63,9 +63,9 @@ export async function render_response({
 
 	const { client } = manifest._;
 
-	const modulepreloads = new Set([...client.start.imports, ...client.app.imports]);
-	const stylesheets = new Set(client.app.stylesheets);
-	const fonts = new Set(client.app.fonts);
+	const modulepreloads = new Set(client.imports);
+	const stylesheets = new Set(client.stylesheets);
+	const fonts = new Set(client.fonts);
 
 	/** @type {Set<string>} */
 	const link_header_preloads = new Set();
@@ -292,10 +292,9 @@ export async function render_response({
 		const blocks = [];
 
 		const properties = [
-			`env: ${s(public_env)}`,
 			paths.assets && `assets: ${s(paths.assets)}`,
 			`base: ${base_expression}`,
-			`element: document.currentScript.parentElement`
+			`env: ${s(public_env)}`
 		].filter(Boolean);
 
 		if (chunks) {
@@ -318,7 +317,9 @@ export async function render_response({
 						${properties.join(',\n\t\t\t\t\t\t')}
 					};`);
 
-		const args = [`app`, `${global}.element`];
+		const args = [`app`, `element`];
+
+		blocks.push(`const element = document.currentScript.parentElement;`);
 
 		if (page_config.ssr) {
 			const serialized = { form: 'null', error: 'null' };
@@ -355,8 +356,8 @@ export async function render_response({
 		}
 
 		blocks.push(`Promise.all([
-						import(${s(prefixed(client.start.file))}),
-						import(${s(prefixed(client.app.file))})
+						import(${s(prefixed(client.start))}),
+						import(${s(prefixed(client.app))})
 					]).then(([kit, app]) => {
 						kit.start(${args.join(', ')});
 					});`);