+++ packages/kit/src/runtime/shared.js
@@ -10,3 +10,5 @@ export function validate_depends(route_id, dep) {
 		);
 	}
 }
+
+export const INVALIDATED_PARAM = 'x-sveltekit-invalidated';