+++ packages/kit/test/apps/basics/test/cross-platform/test.js
@@ -6,43 +6,28 @@ import { test } from '../../../../utils.js';
 test.describe.configure({ mode: 'parallel' });
 
 test.describe('CSS', () => {
-	test('applies imported styles', async ({ page }) => {
+	test('applies imported styles', async ({ page, get_computed_style }) => {
 		await page.goto('/css');
 
-		expect(
-			await page.evaluate(() => {
-				const el = document.querySelector('.styled');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(255, 0, 0)');
+		expect(await get_computed_style('.styled', 'color')).toBe('rgb(255, 0, 0)');
 	});
 
-	test('applies layout styles', async ({ page }) => {
+	test('applies layout styles', async ({ page, get_computed_style }) => {
 		await page.goto('/css');
 
-		expect(
-			await page.evaluate(() => {
-				const el = document.querySelector('footer');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(128, 0, 128)');
+		expect(await get_computed_style('footer', 'color')).toBe('rgb(128, 0, 128)');
 	});
 
-	test('applies local styles', async ({ page }) => {
+	test('applies local styles', async ({ page, get_computed_style }) => {
 		await page.goto('/css');
 
-		expect(
-			await page.evaluate(() => {
-				const el = document.querySelector('.also-styled');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(0, 0, 255)');
+		expect(await get_computed_style('.also-styled', 'color')).toBe('rgb(0, 0, 255)');
 	});
 
-	test('applies imported styles in the correct order', async ({ page }) => {
+	test('applies imported styles in the correct order', async ({ page, get_computed_style }) => {
 		await page.goto('/css');
 
-		const color = await page.$eval('.overridden', (el) => getComputedStyle(el).color);
+		const color = await get_computed_style('.overridden', 'color');
 		expect(color).toBe('rgb(0, 128, 0)');
 	});
 });
@@ -271,7 +256,7 @@ test.describe('Errors', () => {
 		});
 	}
 
-	test('server-side load errors', async ({ page }) => {
+	test('server-side load errors', async ({ page, get_computed_style }) => {
 		await page.goto('/errors/load-server');
 
 		expect(await page.textContent('footer')).toBe('Custom layout');
@@ -279,12 +264,7 @@ test.describe('Errors', () => {
 			'This is your custom error page saying: "Crashing now"'
 		);
 
-		expect(
-			await page.evaluate(() => {
-				const el = document.querySelector('h1');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(255, 0, 0)');
+		expect(await get_computed_style('h1', 'color')).toBe('rgb(255, 0, 0)');
 	});
 
 	test('404', async ({ page }) => {
@@ -595,7 +575,7 @@ test.describe('Redirects', () => {
 			const [, response] = await Promise.all([
 				app.goto('/redirect/in-handle?throw&cookies'),
 				page.waitForResponse((request) =>
-					request.url().endsWith('in-handle/__data.json?throw=&cookies=&x-sveltekit-invalidated=_1')
+					request.url().endsWith('in-handle/__data.json?throw=&cookies=&x-sveltekit-invalidated=01')
 				)
 			]);
 			expect((await response.allHeaders())['set-cookie']).toBeDefined();
@@ -763,19 +743,6 @@ test.describe('Routing', () => {
 		expect(await page.textContent('h1')).toBe('Great success!');
 	});
 
-	test('back button returns to previous route when previous route has been navigated to via hash anchor', async ({
-		page,
-		clicknav
-	}) => {
-		await page.goto('/routing/hashes/a');
-
-		await page.locator('[href="#hash-target"]').click();
-		await clicknav('[href="/routing/hashes/b"]');
-
-		await page.goBack();
-		expect(await page.textContent('h1')).toBe('a');
-	});
-
 	test('focus works if page load has hash', async ({ page, browserName }) => {
 		await page.goto('/routing/hashes/target#p2');
 
@@ -800,23 +767,14 @@ test.describe('Routing', () => {
 	});
 
 	test(':target pseudo-selector works when navigating to a hash on the same page', async ({
-		page
+		page,
+		get_computed_style
 	}) => {
 		await page.goto('/routing/hashes/target#p1');
 
-		expect(
-			await page.evaluate(() => {
-				const el = document.getElementById('p1');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(255, 0, 0)');
+		expect(await get_computed_style('#p1', 'color')).toBe('rgb(255, 0, 0)');
 		await page.click('[href="#p2"]');
-		expect(
-			await page.evaluate(() => {
-				const el = document.getElementById('p2');
-				return el && getComputedStyle(el).color;
-			})
-		).toBe('rgb(255, 0, 0)');
+		expect(await get_computed_style('#p2', 'color')).toBe('rgb(255, 0, 0)');
 	});
 
 	test('last parameter in a segment wins in cases of ambiguity', async ({ page, clicknav }) => {
@@ -962,6 +920,28 @@ test.describe('Routing', () => {
 			expect(await page.textContent('h1')).toBe('symlinked');
 		});
 	}
+
+	test('trailing slash server with config always', async ({ page, clicknav }) => {
+		await page.goto('/routing/trailing-slash-server');
+		await clicknav('[href="/routing/trailing-slash-server/always"]');
+		expect(await page.textContent('[data-test-id="pathname-store"]')).toBe(
+			'/routing/trailing-slash-server/always/'
+		);
+		expect(await page.textContent('[data-test-id="pathname-data"]')).toBe(
+			'/routing/trailing-slash-server/always/'
+		);
+	});
+
+	test('trailing slash server with config never', async ({ page, clicknav }) => {
+		await page.goto('/routing/trailing-slash-server');
+		await clicknav('[href="/routing/trailing-slash-server/never/"]');
+		expect(await page.textContent('[data-test-id="pathname-store"]')).toBe(
+			'/routing/trailing-slash-server/never'
+		);
+		expect(await page.textContent('[data-test-id="pathname-data"]')).toBe(
+			'/routing/trailing-slash-server/never'
+		);
+	});
 });
 
 test.describe('XSS', () => {