+++ packages/kit/test/apps/embed/src/routes/embed/+page.js
@@ -1,0 +1,7 @@
+/** @type {import('@sveltejs/kit').Load} */
+export async function load({ fetch }) {
+	return {
+		a: fetch('/embed/a').then((x) => x.text()),
+		b: fetch('/embed/b').then((x) => x.text())
+	};
+}