+++ packages/adapter-cloudflare/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare
 
+## 2.2.2
+
+### Patch Changes
+
+- fix: respect response status code for caching in `adapter-cloudflare` ([#9820](https://github.com/sveltejs/kit/pull/9820))
+
+## 2.2.1
+
+### Patch Changes
+
+- chore: update all dependencies with minor version bumps ([#9761](https://github.com/sveltejs/kit/pull/9761))
+
 ## 2.2.0
 
 ### Minor Changes